Add Package::cryptographicHash(QCryptographicHash::Algorithm)
This method is intended to replace contentsHash which operates only
on Sha1. In order to support more secure hashing algorithmns and also
to support future developments the new implementation does not hard
code the algorithm but allows to specify it. By that the existing
implementation can just delegate to the new one.

Another change in the implementation is that the new cryptographicHash
method returns a QByteArray instead of a QString. As only a hex
representation of the hash is returned the conversion to QString is
not necessary.

Package::contentsHash() is marked as deprecated.

REVIEW: 126883