Let kpackagetool5 generate appstream information for kpackage components
We want our components to be exposed on software using Appstream (such as
Discover, but not exclusively). Some of the most hidden components we have
nowadays are Plasma addons, since they're not documented there. This
attempts to fix this by reusing the metadata we already have and turning
it into files that appstream can consume, without having to write yet
another time the plugin metadata.

This is done in 2 steps:
* Introduces kpackagetool5 --appstream-metainfo argument
* The cmake macro will automatically grab the file and process it.

REVIEW: 128590