Improve kpackage_install_package
Make sure we don't just append to the regenerateindex.sh file but refresh
it on every configure execution.
Don't call generate every time when we configure. It's bound to break on
systems that install in /usr.
Use the KF5::kpackagetool5 imported target rather than trying to figure out
where it went.

REVIEW: 128599