Generate metadata.json file at build time
In plasma-framework internally we use a QJsonObject for reading the
metadata. Providing the metadata in desktop only makes the system have to
parse metadata.desktop and then convert it internally into a QJsonObject.
Furthermore, to do so we use a desktop file parser that is not very
performant (at least in comparison to KConfig's).

As part of #129102, we made sure that plasma supports metadata.json. This
patch leverages it moving the weight of converting desktop to json at
compile time.

REVIEW: 129103