manually load metadata if absolute path is passed
if the path passed is an absolute path, and a metadata
file is found under it, use that metadata file to
generate the appstream info.
This can happen in the case an application wanting to
support kpackage based extensions includes in the same
project both the packagestructure plugin and the packages
themselves. In that case at build time the
packagestructure plugin wouldn't be installed yet

reviewed-by:Eike Hein <hein@kde.org>