Improve installation process of KPackage bundles
When parsing the desktop files, pass together the kpackage-generic.desktop
service type, so that the parser knows its intrinsic types (most especially
 X-KPackage-Dependencies which is in fact added now). This is important
because otherwise we get warnings when we extract the value when the
package is a desktop file.

This would be much less of a problem if desktop service types could be
extended, but it's not the case at the moment.

REVIEW: 129716