Ignore warnings during appdata generation
Summary:
The Plasma/PopupApplet service type for example got deprecated and removed,
but is still used by various parts of Plasma 5.8.
This causes kpackagetool to emit warnings, but it still succeeds. However,
this case is handled incorrectly as any stderr output is assumed to be critical
and the file is not installed. As result is 0, no warning message is printed.

This commit fixes this by ignoring the stderr output and using result to indicate
success. Warning messages are printed if applicable.

Major issue is that kpackagetool5 does not exit with an error if it encounters
an error, so this installs a broken appdata file in some circumstances.

BUG: 378529

Test Plan:
Built Plasma 5.8 with this patch. Works fine now.
As warnings are now printed, it uncovered some little issues as well.

Reviewers: mart, apol

Reviewed By: mart

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D5328