adopt new KAboutLicense::spdx
Summary:
Previously we'd use the verbatim License key of the kpackage desktop file
as license in appstream. This is however incorrect since appstream expects
SPDX license expressions which use a different format than the one we
use to express KAboutLicenses. This resulted in basically every kpackage
which has an appstream metadata file to cause warnings in our CI and the
distribution generator tool as the license field was using an invalid
format. To resolve this situation a new spdx() method was added to
KAL allowing the conversion from our strings to SPDX compliant expressions.

This additionally tracks and-later-versions qualification (i.e. trailing
plus signs to indicate leter versions than the one defined are acceptable).

To adopt this we now construct a KAL from the keyword string and then
grab the spdx() of that for use as the appstream license.
Also the test desktop file and expectations were adjusted for this and
now include an and-later-versions qualification to make sure that works
as well.

Test Plan: tests pass

Reviewers: sebas, apol

Reviewed By: apol

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D6793