kpackage: bundle a copy of servicetypes/kpackage-generic.desktop
Summary:
This allows to remove the dependency on the installed file, fixing
"make test" in kpackage before installing it.

I extended the unittest to have a list of deps, but it's really in a
.desktop file that it would matter for checking that the servicetype
is correctly found; didn't find a unittest for that though
(both a desktop file and dependencies).

[We could also just get rid of the servicetype desktop file being passed
to KPluginMetadata, AFAICS, given that its properties are just strings
and dependencies are looked up explicitly as a stringlist - but I guess
we want to keep it in case there are more properties in the future]

Test Plan: make uninstall ; ctest .

Reviewers: mart

Subscribers: #frameworks

Differential Revision: https://phabricator.kde.org/D7243