Generate the rcc file every time, at install time
Summary:
In the case of bundles, installing consists of putting all the files in the rcc
and put this file on the right place.
Instead of juggling cmake to remove the file and create it every time, do it
explicitly at installation time.

Test Plan:
Now I can build plasma-desktop, instead of having it build the rcc file and
then remove it before installing

Reviewers: #frameworks, #plasma, mart

Reviewed By: #plasma, mart

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D11582