Fix the rcc binary package generation
Summary:
kpackage_install_bundled_package used install(CODE ..) for the
generation and installation of the rcc file. This had multiple flows,

- It didn't respect the DESTDIR variable of install target which made
most of distribution packaging fail as it tried to write to installation
prefix directly.

- It used the execute_process which resulted in the rcc file generation
at time of install instead of the build-time, so it would fail with
permission errors if make install is run as normal user

Test Plan:
- tested with kdeplasma-addons that it installs all the applets
- it installs the content.rcc in $prefix/share/plasma/plasmoids/$id/

Reviewers: mart, davidedmundson, apol

Reviewed By: davidedmundson

Subscribers: kossebau, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D11642