remove kpackage_install_package deprecation warning
Summary:
an ad-hoc discussion on #plasma-devel suggests the science isn't quite in
on whether RCCs are in fact always better.
on the one hand simple profiling suggested for plasma use cases it may be
ever so slightly faster in cold boot scenario. on the other hand the
actual rcc code supposedly is fairly inefficient in how it access the
disk and manages memory.

this currently puts us in a situation where rcc and !rcc may be viable
depending on the specific scenario of a package. (depending on disk type,
size of rcc, loading, ram constraints, operating system).

if someone wants to see rcc be default it's likely that improvements need
to be made to Qt first. in any event actually getting some solid profiling
before deprecation would be greatly appreciated. is rcc always better?
why is it better? if it isn't always, what are the criteria that make it
preferred over flat files? does the disk type (rotating/!rotating) matter?
etc. etc.

Test Plan:
- cmake && make pass
- plasma-pa no longer raises deprecation warnings

Reviewers: mart

Reviewed By: mart

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D19383