Remove kpluginindex.json caching
As discussed in https://phabricator.kde.org/T14090 caching mechanism is
based in deprecated Qt API. Based on the benchmarks the performance
implications are not critical.

Also the usages in the KCMs will get optimized to not
reload all of the installed packages whenever anything in KNewStuff is
changed, see
https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/410/diffs?commit_id=4afd5a391290bd153b834e04420afea8795e1a4f.
and https://bugs.kde.org/show_bug.cgi?id=431568.