Introduce new KPackageStructure property to identify plugins & packages
Before the pluginId was used, but explicitly defining this property in
the JSON file will be removed in KF6 in favor of the filename.

Also KPackageStructure key makes the terminology more clear, because the
same key can be used inside of the packages instead of ServiceTypes.

Task: https://phabricator.kde.org/T14563