KPackage: Copy KPluginMetaData object when copying d-ptr
This way we avoid pointless reparsing of the metadata files.
In case of desktop files this is even worse, because the conversion
code needs to be run too.

The KPluginMetaData object is already implicitly shared and
only manually reset when the package path is changed.

Task: https://phabricator.kde.org/T14757