Utilize type-safe KPluginMetaData API methods
Reading a bool value is not possible using a KPluginMetaData::value
overload. This takes also care of the string value type.
Having this key inside of the KPlugin json-object can not happen, since
the desktoptojson compiler only puts selected keys there.

Also we can read the description using the accessor function to check if it is
empty.