QueryTest: validate the install() result
This test current fails on macOS (and Windows). It turns out this is
happening because the testpackagesdep plugin can't be installed but that
is rather hard to debug without checking the KJob results. This commit
adds a check that produces a useful error message if the jobs don't
succeed as expected.