PackageJob: Use QStandardPaths::findExecutable() to find resolvers
On windows the resolvers will have a .exe suffix, but we weren't
considering that when checking for the resolvers specified by
KPACKAGE_DEP_RESOLVERS_PATH and always falling back to the default path.
This should hopefully allow tests to find the mock resolver executable (I
don't have a windows system for testing). This also gets us closer to
all tests passing on macOS, but that will need one more commit.