Simplify code using KPluginMetaData::fromJsonFile
This way we do not have to parse the json ourselves and get more precise
logging from KCoreAddons about errors.

Also the MetadataFileMissingError error code did not make sense where it was put in: If there
is not metadata file the entries would be empty. Meaning we would not enter the if-statement at all.