GenericPackage: Remove X-Plasma-MainScript key
Having a Plasma key in KPackage doesn't really make sense due to the
Plasma prefix.

Also, this feature has little actual use and when used, it often causes one to guess/have a look at
the metadata to find out which file is the entrypoint to the package.
For KCMs, this feature was removed when porting them away from KPackage.

For the generic/QML KPackages, this feature seems unused in KF6. I will
also remove it in plasma-frameworks, but there porting is required.