Don't double emit emitResult()
setupNotificationsOnJobFinished connects to jobThreadFinished and in
there emitResult() is called, so avoid a second connection that would
lead to a double emitted signal.

connect to installPathChanged before connecting to jobThreadFinished.
Signal order matters here, as some client code relies that the package
path() to already be the final "correct" one when the job finishes

together with https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/2535

BUG:492557