packagestructure_compat_p: add Icon field
The comic applet in kdeplasma-addons still uses metadata.desktop files.
It uses the Icon field of the .desktop file for comic icons; but this
field is currently unsupported by the
compatibility layer, and as it needs to be part of the KPlugin section
it can't be read it via the extra keys.

This change adds support for this field to the compatibility layer.