fix: don't install random files outside contents subdir
The documentation specifies that the root of a package should contain
only the `metadata.json` file and a `contents` subdirectory.

> The special, main and always required for every package structure file
> is the "metadata" file

> All the other files are under the contents/ subdirectory

We were however installing all files found in the package root, which
could lead to unexpected files being installed outside of the contents
subdirectory.

This commit changes the installation logic to only install the
metadata.json file and the contents/ subdirectory, ignoring any other
files found in the package root, to match what the documentation
specifies as well as common practice.